+++ src/react.ts
@@ -2,6 +2,8 @@
 
 import './styles.css'
 
+export { MatrixAvatar } from './components/MatrixAvatar'
+export type { MatrixAvatarProps } from './components/MatrixAvatar'
 export { Scene } from './components/Scene'
 export { SceneFallback } from './components/SceneFallback'
 export { SceneWebgl } from './components/SceneWebgl'